agregue la ruta events y announcements
