+++ .gitignore
@@ -2,3 +2,4 @@
 npm-debug.log/
 .DS_Store
 package-lock/
+node_modules/
